# Bzip input stream simple vectorization