--- a/src/ICSharpCode.SharpZipLib/BZip2/BZip2InputStream.cs
+++ b/src/ICSharpCode.SharpZipLib/BZip2/BZip2InputStream.cs
@@ -19,7 +19,11 @@ namespace ICSharpCode.SharpZipLib.BZip2
 		private const int NO_RAND_PART_B_STATE = 6;
 		private const int NO_RAND_PART_C_STATE = 7;
 
-		#endregion Constants
+#if NETSTANDARD2_1
+		private static readonly int VectorSize = System.Numerics.Vector<byte>.Count;
+#endif
+
+#endregion Constants
 
 		#region Instance Fields
 
@@ -711,10 +715,22 @@ namespace ICSharpCode.SharpZipLib.BZip2
 					unzftab[seqToUnseq[tmp]]++;
 					ll8[last] = seqToUnseq[tmp];
 
-					for (int j = nextSym - 1; j > 0; --j)
+					var j = nextSym - 1;
+
+#if !NETSTANDARD2_0 && !NETFRAMEWORK
+					while(j >= VectorSize)
 					{
-						yy[j] = yy[j - 1];
+						var arrayPart = new System.Numerics.Vector<byte>(yy, j - VectorSize);
+						arrayPart.CopyTo(yy, j - VectorSize + 1);
+						j -= VectorSize;
 					}
+#endif
+
+					while(j > 0)
+					{
+						yy[j] = yy[--j];
+					}
+
 					yy[0] = tmp;
 
 					if (groupPos == 0)