@Numpsy 2021-04-28
I once had a (very) brief go at using the SSE intrinsics in the deflate code, but never tried just System.Numerics.Vector on it's own (those require a NetCore3+ TFM to build though).
Nice performance gain from an isolated change though :-)