@konrad-kruczynski 2021-04-30
> Nothing wrong with the implementation, but it took a while staring at it to understand what was happening. Perhaps some comments around it?

Sure, I'l add some.

> 
> The speed improvements are really nice and badly needed for the bzip2 algo which is painfully slow (among other problems).

I would also like for some of you to confirm those speed improvements. Just to be sure that my conclusions are correct.