@konrad-kruczynski 2021-04-30
> I once had a (very) brief go at using the SSE intrinsics in the deflate code, but never tried just System.Numerics.Vector on it's own (those require a NetCore3+ TFM to build though).
> Nice performance gain from an isolated change though :-)

Right. The sole purpose of choosing `System.Numerics.Vector` instead of going directly with `SSE` class was portability, e.g. it will work on ARM out of the box. Nonetheless the first version was written using SSE directly.