@Numpsy 2021-04-30
> I would also like for some of you to confirm those speed improvements. Just to be sure that my conclusions are correct.

this is what I get using your branch

```
BenchmarkDotNet=v0.12.1, OS=Windows 10.0.19041.928 (2004/?/20H1)
Intel Core i7-5820K CPU 3.30GHz (Broadwell), 1 CPU, 12 logical and 6 physical cores
.NET Core SDK=5.0.300-preview.21180.15
  [Host]     : .NET Core 3.1.13 (CoreCLR 4.700.21.11102, CoreFX 4.700.21.11602), X64 RyuJIT
  Job-OUVGAT : .NET Core 2.1.26 (CoreCLR 4.6.29812.02, CoreFX 4.6.29812.01), X64 RyuJIT
  Job-MRCUFD : .NET Core 3.1.13 (CoreCLR 4.700.21.11102, CoreFX 4.700.21.11602), X64 RyuJIT
  Job-RAVTER : .NET Framework 4.8 (4.8.4300.0), X64 RyuJIT


|         Method |        Job |     Toolchain |    Mean |    Error |   StdDev | Ratio |
|--------------- |----------- |-------------- |--------:|---------:|---------:|------:|
| DecompressData | Job-OUVGAT | .NET Core 2.1 | 4.850 s | 0.0354 s | 0.0332 s |  1.00 |
| DecompressData | Job-MRCUFD | .NET Core 3.1 | 2.457 s | 0.0063 s | 0.0056 s |  0.51 |
| DecompressData | Job-RAVTER |        net461 | 4.862 s | 0.0411 s | 0.0384 s |  1.00 |
```