@piksel 2021-05-03
Yeah, got basically the same results:
```
BenchmarkDotNet=v0.12.1, OS=Windows 10.0.19042
AMD Ryzen 7 3800X, 1 CPU, 16 logical and 8 physical cores
.NET Core SDK=5.0.300-preview.21180.15
  [Host]     : .NET Core 3.1.14 (CoreCLR 4.700.21.16201, CoreFX 4.700.21.16208), X64 RyuJIT
  Job-NUXXKY : .NET Core 2.1.27 (CoreCLR 4.6.29916.01, CoreFX 4.6.29916.03), X64 RyuJIT
  Job-BNPYUR : .NET Core 3.1.14 (CoreCLR 4.700.21.16201, CoreFX 4.700.21.16208), X64 RyuJIT
  Job-FDMENM : .NET Framework 4.8 (4.8.4300.0), X64 RyuJIT


|         Method |        Job |     Toolchain |    Mean |    Error |   StdDev | Ratio |
|--------------- |----------- |-------------- |--------:|---------:|---------:|------:|
| DecompressData | Job-NUXXKY | .NET Core 2.1 | 3.698 s | 0.0520 s | 0.0461 s |  1.00 |
| DecompressData | Job-BNPYUR | .NET Core 3.1 | 1.900 s | 0.0373 s | 0.0415 s |  0.51 |
| DecompressData | Job-FDMENM |        net461 | 3.702 s | 0.0463 s | 0.0411 s |  1.00 |
```